create_doc_dit_v6/requirements.txt
ipykernel
langchain==0.1.16
langchain-community==0.0.33
langchain-openai==0.1.3
openai==1.55.3  
huggingface_hub==0.22.2
transformers==4.39.3
jinja2==3.1.3
tiktoken==0.6.0
pypdf==4.2.0
yt_dlp==2024.4.9
pydub==0.25.1
beautifulsoup4==4.12.3
ffmpeg
ffprobe
python-dotenv
sentence-transformers==2.7.0
langchain-chroma
faiss-cpu
lark
python-dotenv